+++ src/SimpleMovingAverage.java
@@ -1,3 +1,5 @@
+
+
 /**
  * Unweighted mean of the last N data points
  *
@@ -8,8 +10,8 @@ public class SimpleMovingAverage extends Strategy {
     /**
      * for use by Triangular Moving Averages
      */
-    protected CircularFIFOBuffer<Double> fastSMABuffer = new CircularFIFOBuffer(FAST_PERIOD);
-    protected CircularFIFOBuffer<Double> slowSMABuffer = new CircularFIFOBuffer(FAST_PERIOD);
+    protected CircularFIFOBuffer<Float> fastSMABuffer = new CircularFIFOBuffer(FAST_PERIOD);
+    protected CircularFIFOBuffer<Float> slowSMABuffer = new CircularFIFOBuffer(FAST_PERIOD);
 
     public SimpleMovingAverage() {
         type = "Simple Moving Average";
@@ -19,39 +21,39 @@ public class SimpleMovingAverage extends Strategy {
      * @return
      */
     @Override
-    protected double computeSlowMovingAverage() {
+    protected float computeSlowMovingAverage() {
         int t = slowDataBuffer.size();
         
         //for the first N = SLOW_PERIOD = 20 data points, we simply take the average
         if (SLOW_PERIOD > t) {
-            double sum = 0;
-            for (double datapoint : slowDataBuffer) {
+            float sum = 0;
+            for (float datapoint : slowDataBuffer) {
                 sum += datapoint;
             }
             return (sum / t);
         }
 
         //When calculating successive values, a new value comes into the sum and and old one drops out, meaning full summation each time isn't necessary
-        double slowMovingAverage = currentSlowMovingAverage - oldestSlowDatapoint / SLOW_PERIOD + slowDataBuffer.peekLast() / SLOW_PERIOD;
+        float slowMovingAverage = currentSlowMovingAverage - oldestSlowDatapoint / SLOW_PERIOD + slowDataBuffer.peekLast() / SLOW_PERIOD;
         slowSMABuffer.add(slowMovingAverage);
         return slowMovingAverage;
     }
 
     @Override
-    protected double computeFastMovingAverage() {
+    protected float computeFastMovingAverage() {
         int t = fastDataBuffer.size();
         
         //for the first N = FAST_PERIOD = 20 data points, we simply take the average
         if (FAST_PERIOD > t) {
-            double sum = 0;
-            for (double datapoint : fastDataBuffer) {
+            float sum = 0;
+            for (float datapoint : fastDataBuffer) {
                 sum += datapoint;
             }
             return (sum / t);
         }
         
         //When calculating successive values, a new value comes into the sum and and old one drops out, meaning full summation each time isn't necessary
-        double fastMovingAverage = currentFastMovingAverage - oldestFastDatapoint / FAST_PERIOD + fastDataBuffer.peekLast() / FAST_PERIOD;
+        float fastMovingAverage = currentFastMovingAverage - oldestFastDatapoint / FAST_PERIOD + fastDataBuffer.peekLast() / FAST_PERIOD;
         slowSMABuffer.add(fastMovingAverage);
         return fastMovingAverage;
     }
@@ -59,17 +61,17 @@ public class SimpleMovingAverage extends Strategy {
     /**
      *
      * @param index of the element
-     * @return the double at the index of slowSMABuffer
+     * @return the float at the index of slowSMABuffer
      */
-    public double getSlowBufferValue(int index) {
+    public float getSlowBufferValue(int index) {
         return slowSMABuffer.get(index);
     }
 
     /**
      * @param index of the element
-     * @return the double at the index of fastSMABuffer
+     * @return the float at the index of fastSMABuffer
      */
-    public double getFastBufferValue(int index) {
+    public float getFastBufferValue(int index) {
         return fastSMABuffer.get(index);
     }
 }
